готовый коммит с заданием
