+++ Navigation/ViewController.swift
@@ -13,7 +13,7 @@ class ViewController: UIViewController {
 
        override func viewDidLoad() {
            super.viewDidLoad()
-           
+           title = "Profile"
            
        }
